pyproject.toml
[project]
name = "benchmark-invprob-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "benchopt>=1.7.0",
    "deepinv @ git+https://github.com/deepinv/deepinv.git@main",
    "numpy>=2.2.0",
    "torch>=2.7.0",
    "python-casacore>=3.7.0",
    "astropy>=7.2.0",
    "torchkbnufft>=1.5.0",
]

[dependency-groups]
dev = [
    "black",
    "ruff",
    "sphinx>=9.1.0; python_version >= '3.12'",
    "sphinx-autodoc-typehints>=3.6.2; python_version >= '3.12'",
    "sphinx-gallery>=0.20.0; python_version >= '3.12'",
    "sphinx-rtd-theme>=3.1.0; python_version >= '3.12'",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501"]

[tool.uv.sources]
deepinv = { git = "https://github.com/deepinv/deepinv.git" }

[tool.setuptools]
packages = ["benchmark_utils"]
